Hot fix

+++ ppi/data_utils/polypeptide_featurizers.py
@@ -727,14 +727,14 @@ class PDBBindAtomicBigraphComplexFeaturizer(BaseFeaturizer):
         protein_graph.ndata["node_v"] = protein_coords
         # protein edge features
         protein_graph.edata["edge_s"] = protein_graph.edata["e"]
-        protein_graph.edata["edge_v"] = protein_coords[protein_graph.edges()[0]] - protein_coords[protein_graph.edges()[1]]
+        protein_graph.edata["edge_v"] = protein_coords[protein_graph.edges()[0].long()] - protein_coords[protein_graph.edges()[1].long()]
 
         # ligand node features
         ligand_graph.ndata["node_s"] = ligand_graph.ndata["h"]
         ligand_graph.ndata["node_v"] = ligand_coords
         # ligand edge features
         ligand_graph.edata["edge_s"] = ligand_graph.edata["e"]
-        ligand_graph.edata["edge_v"] = ligand_coords[ligand_graph.edges()[0]] - ligand_coords[ligand_graph.edges()[1]]
+        ligand_graph.edata["edge_v"] = ligand_coords[ligand_graph.edges()[0].long()] - ligand_coords[ligand_graph.edges()[1].long()]
 
         # combine protein and ligand coordinates
         X_ca = torch.cat((protein_coords, ligand_coords), axis=0)
